Tune Some Constants via Environment Variables
Some of the constants in the `balance` package can be set at startup
time via environment variables. With this, you can customize the color
and style of the developer shell, turn on debugging visuals to outline
Canvas widgets, and more.

The parser is at `balance/debug.go` and human readable descriptions
are in the `balance/README.md`